global query limit and started moving data fetching observables into separate file
