--- hosting/src/app/state/searching/effects/search.effects.spec.ts
@@ -1,25 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-import { provideMockActions } from '@ngrx/effects/testing';
-import { Observable } from 'rxjs';
-
-import { SearchEffects } from './search.effects';
-
-describe('SearchEffects', () => {
-  let actions$: Observable<any>;
-  let effects: SearchEffects;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [
-        SearchEffects,
-        provideMockActions(() => actions$)
-      ]
-    });
-
-    effects = TestBed.inject(SearchEffects);
-  });
-
-  it('should be created', () => {
-    expect(effects).toBeTruthy();
-  });
-});
